Fixes for review comments
