+++ src/Types.ts
@@ -83,6 +83,22 @@ export interface ExtendedProjectData {
   users: Member[];
 }
 
+export interface ProjectDocument {
+  id?: string;
+
+  created_at?: string;
+
+  created_by?: string;
+
+  updated_at?: string;
+
+  updated_by?: string;
+
+  project_id: string;
+
+  document_id: string;
+}
+
 export interface Group {
   id: string;
