--- src/pages/[lang]/projects/[project]/assignments/index.astro
@@ -1,49 +1,0 @@
----
-import ProjectLayout from '@layouts/project/ProjectLayout.astro';
-import { createSupabaseServerClient } from '@backend/supabaseServerClient';
-import { getMyProfile } from '@backend/crud';
-import { getLangFromUrl, getTranslations } from '@i18n';
-import { getProjectExtended, listDocumentsInProject } from '@backend/helpers';
-import { ProjectAssignments } from '@apps/project-assignments';
-
-const lang = getLangFromUrl(Astro.url);
-
-const supabase = await createSupabaseServerClient(Astro.request, Astro.cookies);
-if (!supabase) {
-  return Astro.redirect(`/${lang}/sign-in?redirect-to=${Astro.url.pathname}`);
-}
-
-const profile = await getMyProfile(supabase);
-if (profile.error || !profile.data) {
-  const error = await fetch(`${Astro.url}/500`);
-  return new Response(error.body, { headers: error.headers, status: 500 });
-}
-
-const projectId = Astro.params.project;
-
-const project = await getProjectExtended(supabase, projectId as string);
-if (project.error || !project.data) {
-  const error = await fetch(`${Astro.url}/404`);
-  return new Response(error.body, { headers: error.headers, status: 404 });
-}
-
-const documents = await listDocumentsInProject(supabase, projectId as string);
-
-if (documents.error) {
-  const path = Astro.url.protocol + '//' + Astro.url.host;
-  const error = await fetch(`${path}/500`);
-  return new Response(error.body, { headers: error.headers, status: 500 });
-}
-
-const i18n = getTranslations(Astro.request, 'project-assignments');
----
-
-<ProjectLayout active='Assignments' user={profile.data} project={project.data}>
-  <ProjectAssignments
-    client:only
-    i18n={i18n}
-    me={profile.data}
-    project={project.data}
-    documents={documents.data}
-  />
-</ProjectLayout>
